% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_and_boxplot.R
\name{summary_and_boxplot}
\alias{summary_and_boxplot}
\title{summary_and_boxplot: Summarize numeric vector and output boxplot}
\usage{
summary_and_boxplot(num_vec, na.rm = TRUE)
}
\arguments{
\item{num_vec}{a numeric vector used to calculate summary statistics. num_vec is short for numeric_vector.}

\item{na.rm}{logical, remove NA entries when generating summary statistics if TRUE.(NOTE: boxplot will not use NA entries regardless of this parameter).
na.rm is what many other functions call the parameter to remove NA, so I just followed the convention.}
}
\value{
a numeric vector containing number of elements, number of unique elements, min, 1st quartile, median, mean, 3rd quartile,
max, and number of NAs (in this order).
}
\description{
Summarize a numeric vector input and output a numeric vector containing number of elements, number of unique elements, min,
1st quartile, median, mean, 3rd quartile, max, and number of NAs.
This function also outputs a box-and-jitter plot with NA entries removed for visualization purposes.
}
\examples{
summary_and_boxplot(1:10)
summary_and_boxplot(c(1, 2, 3, 4, 5))

}
